Phase 5 commit 1: extract notes/editor/markdown/{inline,line,caret}.ts
Pure helpers from LiveEditor.tsx — no React, no component state.

- markdown/inline.ts: escHtml, kindFromPath, renderInline (image, link,
  bold, italic, code patterns), InlineCtx type.
- markdown/line.ts: ClassifiedLine, LineCtx, classifyLine,
  renderTableCells, lineHtml.
- markdown/caret.ts: CaretPos, getCaretLineIndex, readAllText,
  saveCaret, restoreCaret.

LiveEditor.tsx imports the extracted helpers; ~256 lines removed
(696 → 440). Build clean.
