+++ app/src/notes/LiveEditor.tsx
@@ -4,265 +4,9 @@ import React, {
 import type { LinkIndexEntry } from '../data/types.ts';
 import { folderColor } from './types.ts';
 import { uploadNoteAsset } from '../data/http/notes.http.ts';
-
-// ---- HTML helpers ----
-
-function escHtml(s: string): string {
-  return s.replace(/[&<>"]/g, c =>
-    c === '&' ? '&amp;' : c === '<' ? '&lt;' : c === '>' ? '&gt;' : '&quot;',
-  );
-}
-
-function kindFromPath(
-  href: string, currentFolder: string, linkIndex: LinkIndexEntry[],
-): string {
-  if (!href) return 'broken';
-  let folder = currentFolder;
-  const m = /^\.\.\/([^/]+)\//.exec(href);
-  if (m) { folder = m[1]; }
-  // Resolve href to a repo-relative path for index lookup
-  const target = href.startsWith('../') ? href.slice(3) : `${currentFolder}/${href}`;
-  const found = linkIndex.some(e => e.path === target);
-  if (!found && href.includes('/')) return 'broken';
-  return folder;
-}
-
-function renderInline(
-  text: string,
-  opts: { active: boolean; currentFolder: string; linkIndex: LinkIndexEntry[] },
-): string {
-  const { active, currentFolder, linkIndex } = opts;
-  let out = '';
-  let i = 0;
-  while (i < text.length) {
-    const rest = text.slice(i);
-
-    // Image: ![alt](url)
-    const img = rest.match(/^!\[([^\]\n]*)\]\(([^)\n]+)\)/);
-    if (img) {
-      const src = img[2];
-      const displaySrc = src.startsWith('assets/')
-        ? `/api/notes/${encodeURIComponent(currentFolder)}/${src}`
-        : src;
-      out += `<span class="ml-marker is-hideable">${escHtml(img[0])}</span>`
-        + `<img class="ml-image" src="${escHtml(displaySrc)}" alt="${escHtml(img[1])}" loading="lazy" />`;
-      i += img[0].length;
-      continue;
-    }
-
-    // Link: [text](url)
-    let m: RegExpMatchArray | null = rest.match(/^\[([^\]\n]+)\]\(([^)\n]+)\)/);
-    if (m) {
-      const label = m[1];
-      const href = m[2];
-      const kind = kindFromPath(href, currentFolder, linkIndex);
-      const hrefEsc = escHtml(href);
-      if (active) {
-        out += `<span class="ml-marker is-hideable">[</span>`
-          + `<span class="ml-link kind-${escHtml(kind)}" data-href="${hrefEsc}">${escHtml(label)}</span>`
-          + `<span class="ml-marker is-hideable">](${hrefEsc})</span>`;
-      } else {
-        out += `<span class="ml-marker is-hideable">[</span>`
-          + `<span class="ml-link kind-${escHtml(kind)}" data-href="${hrefEsc}">${escHtml(label)}</span>`
-          + `<span class="ml-marker is-hideable">](${hrefEsc})</span>`;
-      }
-      i += m[0].length;
-      continue;
-    }
-
-    // Bold: **text**
-    m = rest.match(/^\*\*([^*\n]+)\*\*/);
-    if (m) {
-      out += `<span class="ml-marker is-hideable">**</span>`
-        + `<span class="ml-bold">${escHtml(m[1])}</span>`
-        + `<span class="ml-marker is-hideable">**</span>`;
-      i += m[0].length;
-      continue;
-    }
-
-    // Italic: *text*
-    m = rest.match(/^\*([^*\n]+)\*/);
-    if (m && rest[0] === '*' && rest[1] !== '*') {
-      out += `<span class="ml-marker is-hideable">*</span>`
-        + `<span class="ml-italic">${escHtml(m[1])}</span>`
-        + `<span class="ml-marker is-hideable">*</span>`;
-      i += m[0].length;
-      continue;
-    }
-
-    // Inline code: `text`
-    m = rest.match(/^`([^`\n]+)`/);
-    if (m) {
-      out += `<span class="ml-marker is-hideable">\`</span>`
-        + `<span class="ml-code">${escHtml(m[1])}</span>`
-        + `<span class="ml-marker is-hideable">\`</span>`;
-      i += m[0].length;
-      continue;
-    }
-
-    const next = rest.search(/[\[*`]/);
-    const chunk = next === -1 ? rest : rest.slice(0, next || 1);
-    out += escHtml(chunk);
-    i += chunk.length;
-  }
-  return out;
-}
-
-type ClassifiedLine =
-  | { kind: 'heading'; level: number; marker: string; body: string }
-  | { kind: 'bullet'; marker: string; body: string }
-  | { kind: 'quote'; marker: string; body: string }
-  | { kind: 'table-sep'; marker: string; body: string }
-  | { kind: 'table-row'; marker: string; body: string }
-  | { kind: 'blank'; marker: string; body: string }
-  | { kind: 'para'; marker: string; body: string };
-
-function classifyLine(text: string): ClassifiedLine {
-  let m: RegExpMatchArray | null;
-  m = text.match(/^(#{1,6})\s+(.*)$/);
-  if (m) return { kind: 'heading', level: m[1].length, marker: m[1] + ' ', body: m[2] };
-  m = text.match(/^(\s*[-*+]\s+)(.*)$/);
-  if (m) return { kind: 'bullet', marker: m[1], body: m[2] };
-  m = text.match(/^(>\s?)(.*)$/);
-  if (m) return { kind: 'quote', marker: m[1], body: m[2] };
-  // Table separator (| --- | :--- | ---: |) — check before table-row
-  if (/^\|([ \t]*:?-+:?[ \t]*\|)+\s*$/.test(text)) return { kind: 'table-sep', marker: '', body: text };
-  // Table row: starts with | and has at least one more |
-  if (/^\|.+\|/.test(text)) return { kind: 'table-row', marker: '', body: text };
-  if (text.trim() === '') return { kind: 'blank', marker: '', body: '' };
-  return { kind: 'para', marker: '', body: text };
-}
-
-// Renders a table row or sep as pipe+cell spans. Pipes use display:none in CSS
-// but remain in the DOM so textContent === original markdown text.
-function renderTableCells(text: string, ctx: LineCtx): string {
-  const parts = text.split('|');
-  let html = '';
-  for (let i = 0; i < parts.length; i++) {
-    if (i === 0) {
-      if (parts[i]) html += escHtml(parts[i]);
-    } else {
-      html += '<span class="ml-pipe">|</span>';
-      const isTrailingEmpty = i === parts.length - 1 && !parts[i].trim();
-      if (!isTrailingEmpty) {
-        html += `<span class="ml-tcell">${renderInline(parts[i], { active: false, ...ctx })}</span>`;
-      }
-    }
-  }
-  return html;
-}
-
-interface LineCtx { currentFolder: string; linkIndex: LinkIndexEntry[] }
-
-function lineHtml(text: string, active: boolean, ctx: LineCtx): { cls: string; inner: string } {
-  const c = classifyLine(text);
-  let cls = 'ml-line';
-  let inner = '';
-
-  if (c.kind === 'heading') {
-    cls += ` is-h${Math.min(c.level, 3)}`;
-    if (active) cls += ' is-active';
-    inner = `<span class="ml-marker is-hideable">${escHtml(c.marker)}</span>`
-      + renderInline(c.body, { active, ...ctx });
-  } else if (c.kind === 'bullet') {
-    cls += ' is-bullet';
-    if (active) cls += ' is-active';
-    inner = `<span class="ml-marker is-hideable">${escHtml(c.marker)}</span>`
-      + renderInline(c.body, { active, ...ctx });
-  } else if (c.kind === 'quote') {
-    cls += ' is-quote';
-    if (active) cls += ' is-active';
-    inner = `<span class="ml-marker is-hideable">${escHtml(c.marker)}</span>`
-      + renderInline(c.body, { active, ...ctx });
-  } else if (c.kind === 'table-sep') {
-    cls += ' is-table-sep';
-    if (active) { cls += ' is-active'; inner = escHtml(text); }
-    else inner = renderTableCells(text, ctx);
-  } else if (c.kind === 'table-row') {
-    cls += ' is-table-row';
-    if (active) { cls += ' is-active'; inner = escHtml(text); }
-    else inner = renderTableCells(text, ctx);
-  } else if (c.kind === 'blank') {
-    cls += ' is-blank';
-    if (active) cls += ' is-active';
-    inner = '<br>';
-  } else {
-    if (active) cls += ' is-active';
-    inner = renderInline(text, { active, ...ctx });
-  }
-  return { cls, inner };
-}
-
-// ---- Caret helpers ----
-
-function getCaretLineIndex(root: HTMLElement): number {
-  const sel = window.getSelection();
-  if (!sel || sel.rangeCount === 0) return -1;
-  const range = sel.getRangeAt(0);
-  let node: Node | null = range.startContainer;
-  while (node && node !== root) {
-    if (node.nodeType === 1 && (node as Element).classList?.contains('ml-line')) break;
-    node = node.parentNode;
-  }
-  if (!node || node === root) return -1;
-  const lines = Array.from(root.querySelectorAll<HTMLElement>(':scope > .ml-line'));
-  return lines.indexOf(node as HTMLElement);
-}
-
-function readAllText(root: HTMLElement): string {
-  const lines = Array.from(root.querySelectorAll<HTMLElement>(':scope > .ml-line'));
-  return lines.map(el => el.textContent?.replace(/\n/g, '') ?? '').join('\n');
-}
-
-interface CaretPos { lineIndex: number; offset: number }
-
-function saveCaret(root: HTMLElement): CaretPos | null {
-  const sel = window.getSelection();
-  if (!sel || sel.rangeCount === 0) return null;
-  const range = sel.getRangeAt(0);
-  const lines = Array.from(root.querySelectorAll<HTMLElement>(':scope > .ml-line'));
-  for (let i = 0; i < lines.length; i++) {
-    const line = lines[i];
-    if (line.contains(range.startContainer) || line === range.startContainer) {
-      let offset = 0;
-      const walker = document.createTreeWalker(line, NodeFilter.SHOW_TEXT);
-      let n: Node | null;
-      while ((n = walker.nextNode())) {
-        if (n === range.startContainer) { offset += range.startOffset; return { lineIndex: i, offset }; }
-        offset += (n.textContent?.length ?? 0);
-      }
-      return { lineIndex: i, offset };
-    }
-  }
-  return null;
-}
-
-function restoreCaret(root: HTMLElement, saved: CaretPos | null) {
-  if (!saved) return;
-  const lines = Array.from(root.querySelectorAll<HTMLElement>(':scope > .ml-line'));
-  const line = lines[saved.lineIndex];
-  if (!line) return;
-  let remaining = saved.offset;
-  const walker = document.createTreeWalker(line, NodeFilter.SHOW_TEXT);
-  let n: Node | null;
-  while ((n = walker.nextNode())) {
-    const len = n.textContent?.length ?? 0;
-    if (remaining <= len) {
-      const range = document.createRange();
-      range.setStart(n, remaining);
-      range.collapse(true);
-      const sel = window.getSelection();
-      if (sel) { sel.removeAllRanges(); sel.addRange(range); }
-      return;
-    }
-    remaining -= len;
-  }
-  const range = document.createRange();
-  range.selectNodeContents(line);
-  range.collapse(false);
-  const sel = window.getSelection();
-  if (sel) { sel.removeAllRanges(); sel.addRange(range); }
-}
+import { escHtml, kindFromPath } from './editor/markdown/inline.ts';
+import { lineHtml } from './editor/markdown/line.ts';
+import { saveCaret, restoreCaret, getCaretLineIndex, readAllText } from './editor/markdown/caret.ts';
 
 // ---- LiveEditor component ----
